Update du XML

+++ src/eclipse/model/xml8/DeviceItemConverter.java
@@ -33,8 +33,8 @@ public class DeviceItemConverter implements Converter {
 		String name = null;
 		String unit = null;
 		int bitsize = 0;
-		int minValue = 0;
-		int maxValue = 0;
+		double minValue = 0;
+		double maxValue = 0;
 		double resolution = 0;
 		double factor = 0;
 		int offset = 0;
@@ -51,8 +51,8 @@ public class DeviceItemConverter implements Converter {
 			// Set the <deviceItem> properties
 			if ("unit".equals(reader.getNodeName())) unit = reader.getValue();
 			else if ("bitsize".equals(reader.getNodeName())) bitsize = Integer.parseInt(reader.getValue());
-			else if ("minvalue".equals(reader.getNodeName())) minValue = Integer.parseInt(reader.getValue());
-			else if ("maxvalue".equals(reader.getNodeName())) maxValue = Integer.parseInt(reader.getValue());
+			else if ("minvalue".equals(reader.getNodeName())) minValue = Double.parseDouble(reader.getValue());
+			else if ("maxvalue".equals(reader.getNodeName())) maxValue = Double.parseDouble(reader.getValue());
 			else if ("resolution".equals(reader.getNodeName())) resolution = Double.parseDouble(reader.getValue());
 			else if ("factor".equals(reader.getNodeName())) factor = Double.parseDouble(reader.getValue());
 			else if ("offset".equals(reader.getNodeName())) offset = Integer.parseInt(reader.getValue());
